requirements.txt
paho-mqtt>=1.6.1
APScheduler>=3.10.0
psycopg2-binary>=2.9.9
